Version 2.0.4
The fax history indexer was refactored to ensure that it saves the history to ./log instead of MEIPASS. This ensures that already-downloaded faxes are not redownloaded, even after being deleted from the fax inbox. Manual download is not impacted.

The scanner operations were enhanced to auto-select the system scanner if only 1 is present or to allow the user to select from any installed scanner if multiple scanners are present. This reduces user input when scanning documents.